cmmit alterações gerais

+++ api-vendas/src/main/java/com/stellato/vendas/domain/lead/entity/LeadInterface.java
@@ -1,12 +1,10 @@
 package com.stellato.vendas.domain.lead.entity;
 
-import java.math.BigDecimal;
-
 import com.stellato.vendas.domain.shared.enumerated.StatusEnum;
 
 public interface LeadInterface {
-	public BigDecimal getId();
-	public void SetId(BigDecimal id);
+	public Long getId();
+	public void SetId(Long id);
 	public StatusEnum getStatus();
 	public void setStatus(StatusEnum status);
 }
